rename module key_storage -> storage
